"Added git status check to project build process
The project's build scripts `src/project-builder.ts` and `build.mjs` have been updated to include a check for clean git status before continuing the build. This ensures that all changes are committed before a new build is processed, preventing potential loss of uncommitted changes. Furthermore, semver choices are prompted in 'build.mjs' and simpleGit push is initiated after dist package setup."

+++ build.mjs
@@ -2,10 +2,20 @@ import { rimraf } from 'rimraf'
 import { execSync } from 'child_process'
 import tsup from 'tsup'
 import fs from 'node:fs'
+import inquirer from 'inquirer'
+import { simpleGit } from 'simple-git'
 
-await rimraf('dist');
+const git = simpleGit()
+const status = await git.status();
 
-execSync('tsc --project tsconfig.json');
+if (!status.isClean()) {
+  console.error('Commit your changes!')
+  process.exit();
+}
+
+await rimraf('dist')
+
+execSync('tsc --project tsconfig.json')
 
 await tsup.build({
   bundle: true,
@@ -15,13 +25,25 @@ await tsup.build({
   entry: ['src/*'],
 })
 
-const packageJson = JSON.parse(fs.readFileSync('package.json', 'utf8'));
-packageJson.peerDependencies = packageJson.dependencies;
-fs.writeFileSync('package.json', JSON.stringify(packageJson, null, 2) + '\n', 'utf8');
+const { semver } = await inquirer.prompt([
+  {
+    choices: ['patch', 'minor', 'major'],
+    message: 'SemVer',
+    name: 'semver',
+    type: 'list',
+  },
+])
+
+execSync(`npm version ${semver}`)
+
+const packageJson = JSON.parse(fs.readFileSync('package.json', 'utf8'))
+packageJson.peerDependencies = packageJson.dependencies
+fs.writeFileSync('package.json', JSON.stringify(packageJson, null, 2) + '\n', 'utf8')
 
 fs.copyFileSync(
   'package.json',
   'dist/package.json',
 )
 
+await simpleGit().push()
 execSync('cd dist && npm publish --access public && cd ..')
